feat(compliance): replace FedRAMP 20x pilot KSI with CR26 2026.06.24.01
Remove Phase One pilot FedRAMP-20x-KSI-Low (v25.05C) frameworks and
dashboards for AWS/Azure/GCP. Add single FedRAMP-20x-KSI frameworks
(46 indicators) for AWS, Azure, GCP, Kubernetes, and M365 from
FedRAMP/rules@7b9b5ea (info.version 2026.06.24.01), with check maps
ported from the FRMR branch and class-aware notes for varies_by_class.

Breaking: CLI/API compliance_id fedramp_20x_ksi_low_* -> fedramp_20x_ksi_*.
